refactor: move aws api to utils
