+++ tests/test_plan_validation.py
@@ -1,7 +1,3 @@
-import sys
-from pathlib import Path
-
-sys.path.append(str(Path(__file__).parent.parent))
 from external_resources_io.terraform import Action, Plan
 from hooks.validate_plan import RDSPlanValidator
